# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Exakta Laboratories turn taking application",
        "version": "1.0.0",
        "title": "API Turnero",
        "contact": {
            "email": "exakta@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:4000",
    "basePath": "/api",
...
    "paths" : {
...
        "/privilege/{idModule}": {
            "get": {
                "tags": ["Modules-Privileges"],
                "summary" : "Get the privileges of a module",
                "description": "Get the privileges of a module",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idModule",
                        "schema": {
                            "type": "string",
                            "example": "61bbaff53bbe316c78686381"
                        },
                        "description": "IdModule from get data",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/PrivilegeRes"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": [
                                    {
                                        "_id": "61c3924e0f3d03914099a4c9",
                                        "moduleId": "61bbaff53bbe316c78686381",
                                        "area": "Laboratorio",
                                        "privilege": 2,
                                        "createdAt": "2021-12-22T21:02:06.589Z",
                                        "updatedAt": "2021-12-22T21:18:38.579Z",
                                        "__v": 0
                                    }
                                ],
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            }
        },
        "/privilege": {
            "post": {
                "tags": ["Modules-Privileges"],
                "summary" : "Create a privileges of a module",
                "description": "Create a privileges of a module",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "privileges",
                        "description": "Privileges object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PrivilegeReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/PrivilegeRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 201,
                                "body": {
                                    "moduleId": "61bbb1323bbe316c7868638e",
                                    "area": "Imagen",
                                    "privilege": 1
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            }
        },
        "/privilege/{id}": {
            "put": {
                "tags": ["Modules-Privileges"],
                "summary" : "Update a privileges of a module",
                "description": "Update a privileges of a module",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "61f1beb722620191e2a34f34"
                        },
                        "description": "Id of the relation of the privilege"
                    },
                    {
                        "in": "body",
                        "name": "privileges",
                        "description": "Partial privileges object",
                        "schema": {
                            "$ref": "#/definitions/PrivilegeUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/PrivilegeRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": {
                                    "_id": "61c3924e0f3d03914099a4c9",
                                    "moduleId": "61bbaff53bbe316c78686381",
                                    "area": "Imagen",
                                    "privilege": 2,
                                    "createdAt": "2021-12-22T21:02:06.589Z",
                                    "updatedAt": "2021-12-22T21:18:38.579Z",
                                    "__v": 0
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 404,
                                "body": {},
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            },
            "delete": {
                "tags": ["Modules-Privileges"],
                "summary" : "Delete a privileges of a module",
                "description": "Delete a privileges of a module",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "61f1beb722620191e2a34f34"
                        },
                        "description": "Id of the relation of the privilege"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "object",
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": {
                                    "message": "Successfully deleted user."
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 404,
                                "body": {},
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            }
        },
...
    }, 
    "definitions": {
...
        "InvalidRes": {
            "type": "object",
            "properties": {
                "message":  {
                    "type": "string"
                }
            }
        },
...
        "PrivilegeRes": {
            "type": "object",
            "properties": {
                "_id":  {
                    "type": "string"
                },
                "moduleId":  {
                    "type": "string"
                },
                "area":  {
                    "type": "string"
                },
                "privilege":  {
                    "type": "number"
                },
                "createdAt":  {
                    "type": "string"
                },
                "updatedAt":  {
                    "type": "string"
                },
                "__v":  {
                    "type": "number"
                }
            }
        },
        "PrivilegeReq": {
            "type": "object",
            "properties": {
                "moduleId":  {
                    "type": "string"
                },
                "area":  {
                    "type": "string"
                },
                "privilege":  {
                    "type": "number"
                }
            },
            "required": [ 
                "moduleId", 
                "area", 
                "privilege"
            ],
            "example": {
                "moduleId":"61bbb1323bbe316c7868638e",
                "area":"Imagen",
                "privilege":1
            }
        },
        "PrivilegeUpdateReq": {
            "type": "object",
            "properties": {
                "moduleId":  {
                    "type": "string"
                },
                "area":  {
                    "type": "string"
                },
                "privilege":  {
                    "type": "number"
                }
            },
            "example": {
                "privilege":2
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "auth",
            "in": "header"
        }
    }
}